requirements.txt
pandas==2.0.0
numpy==1.24.3
scikit-learn==1.3.0
streamlit==1.24.0
matplotlib==3.7.2
requests==2.31.0
python-dotenv==1.0.0
